refactor bank account service
