+++ src/BankAccount.php
@@ -2,10 +2,7 @@
 
 namespace Timedoor\TmdMidtransIris;
 
-use Timedoor\TmdMidtransIris\Exception\BadRequestException;
-use Timedoor\TmdMidtransIris\Exception\UnauthorizedRequestException;
 use Timedoor\TmdMidtransIris\Models\Bank;
-use Timedoor\TmdMidtransIris\Models\BankAccount as BankAccountModel;
 use Timedoor\TmdMidtransIris\Models\BankAccountValidated;
 use Timedoor\TmdMidtransIris\Utils\Map;
 use Timedoor\TmdMidtransIris\Utils\ConvertException;
@@ -13,57 +10,10 @@ use Timedoor\TmdMidtransIris\Utils\ConvertException;
 class BankAccount extends BaseService
 {
     /**
-     * Get all bank accounts
-     * @throws  UnauthorizedRequestException
-     * @return  BankAccountModel[]
-     */
-    public function all()
-    {
-        $response = $this->apiClient->get('/bank_accounts');
-
-        ConvertException::fromResponse($response);
-
-        $body   = $response->getBody();
-        $result = [];
-
-        if (is_array($body)) {
-            foreach ($body as $item) {
-                $item       = new Map($item);
-                $result[]   = (new BankAccountModel)
-                                ->setId($item->get('bank_account_id'))
-                                ->setBankName($item->get('bank_name'))
-                                ->setAccountName($item->get('account_name'))
-                                ->setAccountNumber($item->get('account_number'))
-                                ->setStatus($item->get('status'));
-            }
-        }
-
-        return $result;
-    }
-
-    /**
-     * Get balance of a certain bank account
-     *
-     * @param   string $id
-     * @throws  UnauthorizedRequestException|BadRequestException
-     * @return  int
-     */
-    public function balance(string $id)
-    {
-        $response = $this->apiClient->get(sprintf('/bank_accounts/%s/balance', $id));
-
-        ConvertException::fromResponse($response);
-        
-        $body = new Map($response->getBody());
-        
-        return $body->get('balance', 0);
-    }
-
-    /**
      * Get available bank list
      *
-     * @throws  UnauthorizedRequestException|BadRequestException
-     * @return  Bank[]
+     * @throws  \Timedoor\TmdMidtransIris\Exception\UnauthorizedRequestException|\Timedoor\TmdMidtransIris\Exception\BadRequestException
+     * @return  \Timedoor\TmdMidtransIris\Models\Bank[]
      */
     public function bankList()
     {
@@ -91,8 +41,8 @@ class BankAccount extends BaseService
      *
      * @param   string $bankCode
      * @param   string $accountNumber
-     * @throws  UnauthorizedRequestException|BadRequestException
-     * @return  BankAccountValidated
+     * @throws  \Timedoor\TmdMidtransIris\Exception\UnauthorizedRequestException|\Timedoor\TmdMidtransIris\Exception\BadRequestException
+     * @return  \Timedoor\TmdMidtransIris\Models\BankAccountValidated
      */
     public function validate(string $bankCode, string $accountNumber)
     {
